python-backend/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# LangChain & LangGraph - compatible versions
langgraph==0.2.45
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-core>=0.3.17,<0.4.0

# Observability
langfuse==2.53.0

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
